requirements.txt
pandas==1.2.0
progressbar2==3.53.1
scikit-learn==0.24.1
gdown==3.12.2
scipy==1.6.2
pymining==0.2
levenshtein==0.12.0
wikipedia==1.4.0
eventlet==0.30.2
nltk==3.4.5
graphene==2.1.8
stanfordnlp==0.1.2
munkres==1.1.4
pycorenlp==0.3.0
gensim==3.8.3
pymagnitude==0.1.143
google==2.0.1
shortuuid==1.0.1
spacy==2.0.18